requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cffi==1.14.3
colorama==0.4.3
cryptography==3.1.1
cx-Oracle==8.0.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-bootstrap-carousel==0.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-nose==1.4.7
django-tinymce4-lite==1.8.0
entrypoints==0.3
gunicorn==20.0.4
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsmin==2.2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.4.0
mysql-connector==2.2.9
mysql-connector-python==8.0.21
nbconvert==5.6.1
nbformat==5.0.7
nose==1.3.7
notebook==6.0.3
numpy==1.19.1
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
py==1.9.0
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.4
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
South==1.0.2
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
